pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "convocast"
version = "1.0.0"
description = "Convert Confluence pages into onboarding podcasts"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "ConvoCast Team"}
]
keywords = ["confluence", "podcast", "onboarding", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.8"
dependencies = [
    "requests>=2.31.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "pydantic>=2.0.0",
    "pyttsx3>=2.90",
    "rich>=13.0.0",
    "lxml>=4.9.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "isort>=5.12.0"
]

[project.scripts]
convocast = "convocast.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["convocast*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true